Report error on bash(1) 3
Some parts of the script don't work on bash(1) 3 (that's what MacOS has).
Suggest installing a newer bash(1) or using another shell, like zsh(1).

This test needs to go on top of the file, since we need to use very
basic shell features.  bash(1) 3 didn't have many of the extensions it
has today, and even using test(1)'s -v is not possible.

Reported-by: Liam Crilly <liam@nginx.com>
Signed-off-by: Alejandro Colomar <alx@nginx.com>